#14 Fix numbers

+++ src/main/java/com/sngular/annotation/processor/mapping/ByteMapping.java
@@ -40,7 +40,7 @@ public class ByteMapping implements TypeMapping<Integer> {
       result = uniformRandomProvider.nextInt(minValue, maxValue);
     } else {
 
-      result = uniformRandomProvider.nextInt(0, Integer.MAX_VALUE);
+      result = uniformRandomProvider.nextInt(0, Byte.MAX_VALUE);
     }
     return result;
   }
